+++ tests/test_utils.py
@@ -14,6 +14,7 @@ from rest_framework_json_api.utils import (
     format_resource_type,
     format_value,
     get_related_resource_type,
+    get_resource_id,
     get_resource_name,
     get_resource_type_from_serializer,
     undo_format_field_name,
@@ -390,6 +391,22 @@ def test_get_resource_type_from_serializer_without_resource_name_raises_error():
         "can not detect 'resource_name' on serializer "
         "'SerializerWithoutResourceName' in module 'tests.test_utils'"
     )
+
+
+@pytest.mark.parametrize(
+    "resource_instance, resource, expected",
+    [
+        (None, None, None),
+        (object(), {}, None),
+        (BasicModel(id=5), None, "5"),
+        (BasicModel(id=9), {}, "9"),
+        (None, {"id": 11}, "11"),
+        (object(), {"pk": 11}, None),
+        (BasicModel(id=6), {"id": 11}, "11"),
+    ],
+)
+def test_get_resource_id(resource_instance, resource, expected):
+    assert get_resource_id(resource_instance, resource) == expected
 
 
 @pytest.mark.parametrize(